# services/drop/openapi.yaml
openapi: 3.1.0
info:
  title: HomeHub Drop API
  version: 2.0.0
  description: |
    Temporary text and file relay behind HomeHub. Browser clients authenticate
    with HomeHub at the edge; Traefik and Control provide a verified,
    service-bound internal identity to Drop.
servers:
  - url: /drop
    description: HomeHub public route
security:
  - homeHubSession: []
tags:
  - name: Items
  - name: Attachments
  - name: System
paths:
  /health/live:
    get:
      tags: [System]
      summary: Process liveness
      security: []
      responses:
        "204": {description: Alive}
  /health/ready:
    get:
      tags: [System]
      summary: Storage readiness
      security: []
      responses:
        "204": {description: Ready}
        "503": {description: Not ready}
  /api/v1/items:
    get:
      tags: [Items]
      summary: List active items
      parameters:
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 200, default: 50}}
        - {name: cursor, in: query, schema: {type: string}}
      responses:
        "200":
          description: Active items
          content:
            application/json:
              schema:
                type: object
                required: [items, next_cursor]
                properties:
                  items: {type: array, items: {$ref: "#/components/schemas/Item"}}
                  next_cursor: {type: string}
        "401": {$ref: "#/components/responses/Unauthorized"}
    post:
      tags: [Items]
      summary: Create a temporary item
      parameters:
        - {name: Idempotency-Key, in: header, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                text: {type: string}
                ttl_days: {type: integer, enum: [1, 3, 7]}
                files: {type: array, maxItems: 10, items: {type: string, format: binary}}
      responses:
        "201": {description: Created}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
  /api/v1/items/{id}:
    parameters:
      - {$ref: "#/components/parameters/ItemID"}
    get:
      tags: [Items]
      summary: Get item metadata
      responses:
        "200":
          description: Item
          content: {application/json: {schema: {$ref: "#/components/schemas/Item"}}}
        "404": {$ref: "#/components/responses/NotFound"}
    delete:
      tags: [Items]
      summary: Delete an item (owner only)
      responses:
        "204": {description: Deleted}
        "403": {$ref: "#/components/responses/Forbidden"}
  /api/v1/items/{id}/text:
    get:
      tags: [Items]
      summary: Read full item text
      parameters:
        - {$ref: "#/components/parameters/ItemID"}
      responses:
        "200": {description: Text content, content: {text/plain: {schema: {type: string}}}}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/v1/items/{id}/expiry:
    patch:
      tags: [Items]
      summary: Change expiry (owner only)
      parameters:
        - {$ref: "#/components/parameters/ItemID"}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ttl_days]
              properties: {ttl_days: {type: integer, enum: [1, 3, 7]}}
      responses:
        "200": {description: Updated}
        "403": {$ref: "#/components/responses/Forbidden"}
  /api/v1/attachments/{id}:
    get:
      tags: [Attachments]
      summary: Download or display an attachment
      parameters:
        - {$ref: "#/components/parameters/AttachmentID"}
      responses:
        "200": {description: Attachment bytes}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/v1/attachments/{id}/preview:
    get:
      tags: [Attachments]
      summary: Get an image preview
      parameters:
        - {$ref: "#/components/parameters/AttachmentID"}
      responses:
        "200": {description: Preview image}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/v1/status:
    get:
      tags: [System]
      summary: Storage and traffic status (owner only)
      responses:
        "200": {description: Status report}
        "403": {$ref: "#/components/responses/Forbidden"}
  /api/v1/events:
    get:
      tags: [Items]
      summary: Stream item changes
      responses:
        "200": {description: Event stream, content: {text/event-stream: {schema: {type: string}}}}
components:
  securitySchemes:
    homeHubSession:
      type: apiKey
      in: cookie
      name: homehub_session
      description: HomeHub edge session; Drop itself receives only a signed internal identity.
  parameters:
    ItemID: {name: id, in: path, required: true, schema: {type: string}}
    AttachmentID: {name: id, in: path, required: true, schema: {type: string}}
  schemas:
    Attachment:
      type: object
      required: [id, original_name, mime_type, size, previewable, download_url]
      properties:
        id: {type: string}
        original_name: {type: string}
        mime_type: {type: string}
        size: {type: integer, format: int64}
        previewable: {type: boolean}
        download_url: {type: string}
        preview_url: {type: string}
    Item:
      type: object
      required: [id, text_truncated, text_size, has_text, source, created_at, expires_at, total_size, attachments]
      properties:
        id: {type: string}
        text_preview: {type: string}
        text_truncated: {type: boolean}
        text_size: {type: integer, format: int64}
        has_text: {type: boolean}
        source: {type: string, enum: [owner, guest]}
        created_at: {type: string, format: date-time}
        expires_at: {type: string, format: date-time}
        total_size: {type: integer, format: int64}
        attachments: {type: array, items: {$ref: "#/components/schemas/Attachment"}}
        full_text_url: {type: string}
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: {type: string}
            message: {type: string}
  responses:
    BadRequest: {description: Invalid request, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
    Unauthorized: {description: Authentication required, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
    Forbidden: {description: Owner permission required, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
    NotFound: {description: Resource not found, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}